Turned morpho.color.heatmap into a function
morpho.color.heatmap is now a function (instead of a constant) that generates a new heatmap Gradient object when called with no arguments. The purpose is to make it so that if multiple heatmap Gradients are assigned, they will be different objects and won't influence each other if one of them is mutated. It also aligns its usage with that of rainbow().