Refactored `MultiFigure.__setattr__()`
The special setattr of MultiFigure now checks for a native attribute of self by attempting to access it using the base `Figure` class's getattr methods inside a try clause, instead of checking for the attribute in `dir(self)` or `self._state`. Basically, it now does something like `Figure.__hasattr__(self, name)` to check for attribute existence.

This change was made to make this check more standardized and adaptable in case a change is made to the `Figure` class which adds abstract attributes other than tweenables.