Revert "Setting `zdepth` within `paragraph()` now works."
This reverts commit bc924d4059cb57fc8ed863bcb8258d4af6dfb68e.

I decided against it after all, because if I allow `zdepth` to be set this way, then I also have to allow `transition` and `tweenMethod` and all the others as well. I think it's better to leave all of those out of kwargs of `paragraph()` and just make them settable afterward, similar to how most figures don't let you set `zdepth` as an input to the constructor.