Paths no longer support singleton subscripting
This was done since slicing returns a subpath, but singleton subscripts return a complex number, which feels like an inconsistent type convention since one returns a figure, and the other a number.

Plus, it's tempting to confuse `path[n]` with `path.seq[n]` which is completely different, so for now, I'm removing the feature until/unless I get convinced otherwise.