Singleton path slices are now allowed
Slicing a path with `mypath[t]`, or equivalently `mypath[t:t]`, now returns a single node path instead of throwing an error.

I decided this case should be handled since reversed slices are now allowed and because the user might want to create a skit with variable slice endpoint ranging from behind to ahead of the startpoint.