Fixed bug with locator layers
Fixed a bug with locator layers where clicking the screen after the animation has stopped at its final frame would capture the location of the click according to the *first* frame of the animation when replayed. This could cause issues if the starting and ending frames of the animation had differing camera views.