Fixed delay drifts when changing animation speed
Changing the speed of an animation (e.g. by using `newFrameRate()`) could sometimes cause later events in the changed animation to drift out of sync with the original animation. This was due to delay durations being locally rounded with no regard to the global timeline causing round-off errors to accumulate in unlucky cases.