Improved `@SkitParameters()`
The modified constructor `SkitParameters()` creates no longer uses the `update()` method to setup the tweenables. I believe this is an overall better design, and it also allows one to create "hidden" tweenables/nontweenables by setting them up in a separate `__init__()` method defined in the Skit classdef. These are "hidden" in the sense that they cannot be initialized by passing keyword arguments to the constructor, and can only be modified after construction.