Fixed bug with outlined dashed paths with tipped tails
The dash of a path's outline would fall out of sync with the base path if it had a non-zero `tailSize`.

This also introduced a new tweenable to the `Path` class: `dashOffset`. For now, it's just an implementation detail, but later on, it should be made standard for all figures that support a `dash` tweenable.