Loosened behavior of `tweenMethod` decorator
`@tweenMethod` no longer checks for type-matching since tweening is almost always performed at the Actor level which already enforces type matching.

It also no longer forces the tweened figure to inherit the visibility of the first keyfigure. This was done because it didn't enforce any of the other meta-settings (e.g. `static`), and I think it makes more sense for it to be a convention that needs to be manually abided by when creating custom tweens instead of a standard enforced by the decorator.