Removed references to `morpho.array()`
I think `np.array()` is now safe to use anywhere in Morpho, and I don't think it's that bad in practice if users just use vanilla `np.array()` which is probably what they're used to. The main thing is to avoid doing in-place operations on arrays that are non-float and non-complex (e.g. arrays with dtype=int) since the typecasting can mess up the math. But if you mostly treat arrays as if they're immutable, it should be fine.