Fixed bug with `_updateFrom(copy=False)`
Setting `copy=False` used to mean the `Tweenable` object itself of the target figure was directly reassigned to self's `_state` dict, resulting in them being linked even for immutable tweenable values. Now if `copy=False` is set, the `Tweenable` object itself will be *shallow* copied, meaning the `Tweenable` object itself is copied, but the payload value it contains is NOT copied.