Improved typecasting for multiselection attribute gets
When attempting to get a common subfigure attribute value via `.all`, if the common value is a list or tuple, it will use as the return type whatever the type of the initial subfigure is.

Also, if the common value is a np.array, it will be copied before it is returned.