Implemented substaggering and `subaction` feature
The basic actor actions for Frames and MultiFigures now support a `substagger` option that allows the effect of the actions to be "staggered" amongst the subfigures. Note that some MultiFigures are currently excluded (FancyMultiText, notably).

Also implemented the `subaction` actor action that applies actions to the subfigures of a Frame-like Actor along with optional substaggering.