`fadeOut()` now preserves subfigures' static state
Also true of MathGrid `shrinkOut()`. After the action completes, it will restore the static attribute of all subfigures to what they were before the action completed. The thought was this could be useful if the user (for some reason) wanted to reuse the exited actor, but I now kind of think that exiting actions should be understood to somewhat permanently cause an actor to be finished, and the user should not try to do anything more with them after they have exited.