**Box gadgets now return relative paths by default**
`enbox()`, `crossout()` and `encircle()` now return path actors whose paths are positioned so that their `origin` attribute is the center of the box. This makes it much easier to modify the resulting path's transformation tweenables in a way that makes sense.

A new optional keyword parameter `align` has been added to each of these which defaults to `(0,0)` but can be given any alignment setting or set to `None` to use absolute coordinates like how these gadgets used to work.