`fromsvg()` takes stroke widths into account better
Before, any stroke widths found in the SVG were assigned as pixel widths for the splines, but in reality, SVG stroke widths should be interpreted as PHYSICAL widths, so a physical-to-pixel conversion is needed to import an SVG fully faithfully. Note that it's still imperfect in that resizing the spline (or zooming the camera) after the SVG is imported will cause the stroke widths to no longer match the original since pixel widths don't rescale with the figure/scene.