**Changed how subpool works, slightly**
The subpool is no longer stored as a set of literal indices specific to that MultiFigure. Instead, it stores a list of indices, slices, filters, etc. and negative indices are stored as negative so that if the figure list changes, the negative index's target will update automatically. Note that since the subpool is saved as a list, it no longer can be modified with set-specific methods. So this is a breaking change.