Disabled substaggering for FancyMultiText fade actions
It turns out doing a substaggered fade jump behaves incorrectly on certain FancyMultiText figures. I believe the issue has to do with how alignment is handled by reference to the total bounding box of the Fancy figure, but it doesn't seem easy to fix currently, so for now I'm disabling it. Substaggering can still be performed by using the `subaction` command directly, but note that the result may be off.