Refactored `TFrame.combine()`
It now checks for being an instance of TFrame instead of checking for the existence of a `commitTransforms` attribute. I think this is a safer implementation (albeit slightly more restrictive). Plus, for MultiFigures, `commitTransforms` may pass the `hasattr()` test by passing to subfigures via `all` (true, this could still be bypassed by using `object_hasattr()`, but I think it's better to not rely on attribute existence as the check in this case).