Refactored [Multi]Image classes
`Image` now inherits from PreAlignableFigure so that it has access to `alignOrigin()`.

`MultiImage` now inherits from FancyFrame instead of TFrame. This was primarily to allow `alignOrigin()` to work on singleton MultiImages the same as on Images, but it also seems like it might be a good thing to try to do in of itself.