Refactored [Multi][P]Text
`Text` now inherits from PreAlignableFigure so that it has access to `alignOrigin()`.

`MultiText` now inherits from FancyFrame instead of TFrame. This was primarily to allow `alignOrigin()` to work on singleton MultiText figures the same as on Text figures, but it also seems like it might be a good thing to try to do in of itself.