FIXED BUG with paragraph raw boxes
They would remain unchanged after changes of the `align` attribute. Raw boxes are only supposed to stay constant under standard transformation attributes, not `align`. However, it turns out big pieces of code actually relied on that incorrect behavior, so the code had to be refactored to account for the actual correct behavior of raw boxes.