FIXED BUG with `EllipticalArc` rendering
EllipticalArc draw direction is now always consistent with the orientation of `theta0` and `theta1`. Specifically, if `theta1 < theta0`, the arc will be drawn clockwise instead of internally swapping `theta0` and `theta1` to get an equivalent counter-clockwise arc. This is important because the difference is discernible when there's a dash pattern.