**Changed line spacing behavior for paragraphs**
Vertical spacing is now based upon a calculated line height instead of the particular row height for the given line of text. This should make spacing more consistent overall, but it makes ybuf much more sensitive now, so lower values are typically needed to achieve the same kind of line spacing as before.