Made line height alphabet overridable
The alphabet of glyphs used to determine line heights in paragraphs can now be specified as an input to the `paragraph()` function. An alternative default alphabet can also be specified by setting the `LINE_HEIGHT_ALPHABET` constant.
Note that setting the alphabet to be the dollar sign "$" seems to restore line spacings and ybuf calibration pretty close to its old behavior.