FIXED BUG with `popIn/Out()`
Would still show a background box for certain fully shrunk figures if `backPad` was nonzero.

Fixed by disabling the custom `popIn/Out()` actions for `MultiPathBase` and instead using the ones inherited from `FancyFrame`. Note that this could cause issues if ultimately making a truly 3D `SpaceMultiPath` in the future in which case a custom implementation of these actions would probably be required.