Subpath loops are now drawn with closure
This expands the behavior of performing a cairo closure on a drawn path/spline if it ends where it begins to cases where the path/spline contains deadends. It now performs these closures on a per-subpath basis if the subpath ends where it began. Note that this behavior will only activate if the path/spline is full-length (start=0, end=1) AND is arrowless (headSize==tailSize==0).
This change makes PText figures that were turned into splines look nicer when stroked.